making load local input to hdfs optional

+++ importers/src/main/bash/create-presplited-table.sh
@@ -7,6 +7,7 @@ OUTPUT_PATH=$3
 TABLE_NAME=$4
 TABLE_COLFAMS=$5
 TABLE_REGION_CNT=$6
+LOAD_INPUT_TO_HDFS=$7
 
 # global parameters
 COMMONS_JAR=../../../../commons/target/commons-1.0-SNAPSHOT.jar
@@ -14,10 +15,13 @@ HDFS_KEYS_FILE=keys
 
 export HBASE_CLASSPATH=${COMMONS_JAR}
 
+if [ "${LOAD_INPUT_TO_HDFS}" = "true" ]; then
+	hadoop fs -rm -r ${HDFS_INPUT_PATH}
+	hadoop fs -mkdir ${HDFS_INPUT_PATH}
+	hadoop fs -put ${LOCAL_INPUT_PATH} ${HDFS_INPUT_PATH}
+fi
+
 hadoop fs -rm -r ${OUTPUT_PATH}
-hadoop fs -rm -r ${HDFS_INPUT_PATH}
-hadoop fs -mkdir ${HDFS_INPUT_PATH}
-hadoop fs -put ${LOCAL_INPUT_PATH} ${HDFS_INPUT_PATH}
 
 time hadoop jar ${COMMONS_JAR} pl.edu.icm.coansys.commons.hbase.SequenceFileKeysSamplerMR -libjars ${COMMONS_JAR} -D sampler.samples.region.count=${TABLE_REGION_CNT} ${HDFS_INPUT_PATH} ${OUTPUT_PATH}
